fix: close SQL injection across e107 core via e_db bound parameters
Rewrite SQL sinks flagged by a full-core injection audit to the e_db
prepared-statement API, eliminating string interpolation of untrusted
data. Covers 30 CRITICAL, 58 HIGH, and 97 MEDIUM findings.

Techniques:
- createQueryBuilder() fluent builder for the bulk of the rewritten
  sinks: every value is bound through createNamedParameter() and every
  identifier is validated fail-closed, so the call sites read as intent
  (where()/whereIn()/set()/setExpression()/values()/count()) rather
  than hand-assembled SQL.
- execute() with :named bound parameters where the builder cannot model
  the query: raw SQL functions (REGEXP, FIND_IN_SET, SUBSTRING_INDEX,
  TRIM/REPLACE), INSERT ... SELECT, and the foreign-DB connections.
- quoteIdentifier() and strict [A-Za-z0-9_] allowlists for dynamic table
  and column identifiers, which cannot be bound.
- (int) casts for LIMIT/OFFSET and numeric columns.
- Import providers fixed at the root in base_import_class::database()
  (backtick-doubling of the database name plus prefix sanitisation),
  which neutralises all per-provider sinks at once.
- DB abstraction layer (e_db_mysql) hardened at the primitive level with
  identifier validation on select/count/delete/fields and charset and
  language allowlists on SET NAMES, where execute() cannot apply.
- alt_auth otherdb/e107db login fixed via a dedicated secondary e_db
  connection with bound parameters (advisory GHSA-5f75-c25x-59wx).

Moving the faqs add-FAQ INSERT to a named-column values() builder call
also corrects a latent column drift: the old positional INSERT supplied
8 values for the 10-column faqs table, mis-assigning the order count to
faq_author_ip (and erroring outright on a current schema).

The documented "trusted SQL fragment" public APIs in e_db_common_trait
(selectTree/copyRow/max $where, legacy string-form insert, array-form
update WHERE) are left unchanged by contract; their injectable
identifier positions are validated in e_db_mysql instead.

Adds e_db_mysql identifier and charset security tests. Unit suite green.

Second audit pass (independent re-scan) added three sinks the first pass
did not flag, fixed with the same approach:
- alt_auth admin config (alt_auth_adminmenu.php): the option-save loop
  interpolated attacker-controlled $_POST key NAMES into select/update/
  insert; rewritten to the query builder with bound parameters.
- menumanager menuDeactivate: menu_id taken from $_POST['removeid'] was
  not cast before reaching gen/delete/update; now (int).
- user mailout: the last_visit_date selector slipped past a loose numeric
  guard into an unquoted comparison; now (int).
All three require an authenticated admin (or CSRF against one).

+++ e107_plugins/download/handlers/download_class.php
@@ -116,7 +116,7 @@ class download
 			$this->qry['id']		= intval($_GET['id']);
 			$this->qry['order'] 	= vartrue($_GET['order']) && in_array("download_".$_GET['order'],$this->orderOptions) ? $_GET['order'] : $this->qry['order'];
 			$this->qry['sort'] 		= (varset($_GET['sort']) == 'asc') ? "asc" : 'desc';	
-			$this->qry['from']		= varset($_GET['from'],0);
+			$this->qry['from']		= (int) varset($_GET['from'],0);
 
 			if($this->qry['action'] == 'error')
 			{
